ApiHTMLFront/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-formtools==2.5.1
django-oauth-toolkit==2.3.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-two-factor-auth==1.16.0
djangorestframework==3.15.1
drf-jwt==1.19.2
filelock==3.13.3
idna==3.6
jwcrypto==1.5.6
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
phonenumbers==8.13.34
pillow==10.3.0
platformdirs==4.2.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.3.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.1
pyotp==2.9.0